move validate_organization to utils
